refactor(functional): split scanner.rs into statements, expressions, types
